refactor: viewmodel 패키지 정리
