+++ app/src/main/java/woowacourse/shopping/presentation/ui/products/ProductsItemViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.products
+package woowacourse.shopping.presentation.ui.products
 
 import androidx.annotation.CallSuper
 import androidx.databinding.ViewDataBinding
