+++ app/src/main/res/layout/activity_products.xml
@@ -7,7 +7,7 @@
         android:id="@+id/cl_products"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.products.ProductsActivity">
+        tools:context=".presentation.ui.products.ProductsActivity">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_products"
