+++ app/src/main/res/layout/item_load_more.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="onClickHandler"
-            type="woowacourse.shopping.ui.products.LoadMoreViewHolder.OnClickHandler" />
+            type="woowacourse.shopping.presentation.ui.products.LoadMoreViewHolder.OnClickHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
