+++ app/src/test/java/woowacourse/shopping/ui/viewmodel/CartViewModelTest.kt
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.data.dummyProducts
-import woowacourse.shopping.ui.cart.CartViewModel
+import woowacourse.shopping.presentation.viewmodel.cart.CartViewModel
 import woowacourse.shopping.util.InstantTaskExecutorExtension
 import woowacourse.shopping.util.getOrAwaitValue
