+++ app/src/test/java/woowacourse/shopping/ui/viewmodel/ProductsViewModelTest.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import woowacourse.shopping.ui.products.ProductsViewModel
+import woowacourse.shopping.presentation.viewmodel.products.ProductsViewModel
 import woowacourse.shopping.util.InstantTaskExecutorExtension
 import woowacourse.shopping.util.getOrAwaitValue
